requirements.txt
-i https://pypi.org/simple
click==8.1.7; python_version >= '3.7'
numpy==1.26.2; python_version >= '3.10'
opencv-python==4.8.1.78; python_version >= '3.6'
pillow==10.1.0; python_version >= '3.8'
